% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oauth1_signature}
\alias{oauth1_signature}
\title{Generates OAuth1 signature needed to authorize API request}
\usage{
oauth1_signature(method, url, api, params)
}
\arguments{
\item{method}{f.x. "HMAC-SHA1"}

\item{url}{f.x. "https://api.twitter.com/1.1/friends/ids.json"}

\item{api}{f.x. c(cursor=-1, screen_name="hrw", count=10)}

\item{params}{named vector needed for generating oauth1 signature}
}
\value{
signature string
}
\description{
Generates OAuth1 signature needed to authorize API request
}

