% Generated by roxygen2 (4.0.1): do not edit by hand
\name{twitter_api_call}
\alias{twitter_api_call}
\title{GET requests Twitter API}
\usage{
twitter_api_call(url, api, params, print_result = FALSE, use_cygwin = FALSE,
  cygwin_bash = "c:\\\\cygwin64\\\\bin\\\\bash.exe")
}
\arguments{
\item{url}{f.x. "https://api.twitter.com/1.1/friends/ids.json"}

\item{api}{f.x. c(cursor=-1, screen_name="hrw", count=10)}

\item{params}{named vector needed for generating oauth1 signature}

\item{print_result}{TRUE if you want to have resulting JSON printed out}

\item{use_cygwin}{TRUE if you would like to resort to Cygwin/Curl workaround}

\item{cygwin_bash}{f.x. "c:\\cygwin64\\bin\\bash.exe"}
}
\value{
JSON
}
\description{
GET requests Twitter API
}

